Two suggested adjustments to Exercises 14.3 & 14.4
Exercise 14.3 asks about the total population in September 2001 but the exercise uses the 2010 census data to calculate population data. The 2000 census data was loaded in Section 5.2, so suggested a note to use that data and an amended result. 

Exercise 14.4 result was different and after a few checks, propose updated result.