Clarification about this page
All of my modification are for the part "5.2. Loading with ogr2ogr"

As a beginner in every programming languages, some things were for me unclear, so I have decided to propose changes so that it can become more clear for beginners.

I added :
- line 59 (Be sure to be in the :file:`2000/` sub-directory with the shell when doing these instructions:),
- line 67  (Note that on Windows, you will need to use ``set`` instead of ``export``.)
- line 82 (For more visual clarity, these lines are displayed with ``\``, but they should be written in one line on your shell.)

I did some modification on line 47 ( accessible via QGIS4W Shell menu --> accessible via OSGeo4W Shell).

I separate the two parts of export ..... and ogr2ogr because (always as a beginner), I thought that that was one script that should be written in the pgAdmin Query tool (as it was not clearly said that I should open ogr2ogr after installing).

I moved the explanation of the export PGPASSWORD before the running of ogr2ogr -nln ....... so that it can fit with the separation I explained before.